requirements.txt
black==23.1.0
click==8.1.3
keyboard==0.13.5
MouseInfo==0.1.3
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
python3-xlib==0.15
pytweening==1.0.4
tomli==2.0.1
